+++ app/src/main/java/team/retum/jobisandroidv2/ui/BottomMenu.kt
@@ -5,7 +5,7 @@ import androidx.annotation.StringRes
 import team.retum.home.navigation.NAVIGATION_HOME
 import team.retum.jobis.recruitment.navigation.NAVIGATION_RECRUITMENTS
 import team.retum.jobisandroidv2.R
-import team.returm.mypage.navigation.NAVIGATION_MY_PAGE
+import team.retum.jobis.navigation.NAVIGATION_MY_PAGE
 
 sealed class BottomMenu(
     val route: String,
