+++ feature/mypage/src/main/java/team/retum/jobis/viewmodel/MyPageViewModel.kt
@@ -1,4 +1,4 @@
-package team.returm.mypage.viewmodel
+package team.retum.jobis.viewmodel
 
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
